refactor(nitro-colors): move RoleResetCron to a separate file
