+++ src/load-cron.ts
@@ -1,7 +1,7 @@
 import { ActivityCleanCron } from "./programs/activity/controller/activity-clean.cron";
 import { Client } from "discord.js";
 import { RescueForceClose } from "./programs/buddy-project/find-buddy/rescue-force-close";
-import { RoleResetCron } from "./programs/nitro-colors";
+import { RoleResetCron } from "./programs/nitro-colors/roles-reset";
 
 export class LoadCron {
   private static loadCronInstance: LoadCron;
