+++ src/programs/nitro-colors/index.ts
@@ -1,4 +1,4 @@
-import bot from "../index";
+import bot from "../../index";
 import {
   Client,
   Collection,
@@ -11,16 +11,18 @@ import {
   TextChannel,
   User,
 } from "discord.js";
-import { createYesBotLogger } from "../log";
-import { Command, CommandHandler, DiscordEvent } from "../event-distribution";
-import cron from "node-cron";
-import prisma from "../prisma";
-import Tools from "../common/tools";
+import { createYesBotLogger } from "../../log";
+import {
+  Command,
+  CommandHandler,
+  DiscordEvent,
+} from "../../event-distribution";
+import prisma from "../../prisma";
 
-const logger = createYesBotLogger("programs", "NitroColors");
+export const logger = createYesBotLogger("programs", "NitroColors");
 
-let nitroRolesCache: Collection<Snowflake, Role>;
-let colorSelectionMessage: Message;
+export let nitroRolesCache: Collection<Snowflake, Role>;
+export let colorSelectionMessage: Message;
 
 @Command({
   event: DiscordEvent.READY,
@@ -141,69 +143,6 @@ class NitroColorSelector implements CommandHandler<DiscordEvent.REACTION_ADD> {
   }
 }
 
-export class RoleResetCron {
-  static init() {
-    const CRON =
-      process.env.NODE_ENV === "development" ? "* * * * *" : "0 0 1 * *";
-
-    // Schedule a cron task every month
-    cron.schedule(CRON, async () => {
-      // Remove color roles from each user
-      for (const role of nitroRolesCache.values()) {
-        for (const member of role.members.values()) {
-          await member.roles.remove(role);
-        }
-      }
-
-      // Clean up pick-your-color messages
-      const channel = bot.guilds
-        .resolve(process.env.GUILD_ID)
-        ?.channels.cache.find(
-          (c) => c.name === "pick-your-color"
-        ) as TextChannel;
-      const nitroBoosterRole = Tools.getRoleByName(
-        "Nitro Booster",
-        channel.guild
-      );
-
-      if (!nitroBoosterRole) {
-        logger.error("Could not find Nitro Booster role!");
-        return;
-      }
-
-      // Remove all messages sent by the bot
-      const messages = await channel.messages.fetch({ limit: 5 });
-      for (const message of messages.values()) {
-        if (message.id !== colorSelectionMessage.id) {
-          await message.delete();
-          continue;
-        }
-
-        // Remove all reactions from the reactions message
-        const reactionsCache = colorSelectionMessage.reactions.cache;
-        for (const reaction of reactionsCache.values()) {
-          const reactionUsers = await reaction.users.fetch();
-          const usersToRemove = reactionUsers.filter(
-            (user) => user.id !== bot.user?.id
-          );
-
-          for (const [userId] of usersToRemove) {
-            await reaction.users.remove(userId);
-          }
-        }
-      }
-
-      // Let Nitro boosters know about the new month's change!
-      await channel.send({
-        content: `${nitroBoosterRole} It is time to pick a new color for the new month!`,
-      });
-
-      logger.debug("Executed cleanup");
-    });
-
-    logger.debug("Initialized!");
-  }
-}
 const memberHasNitroColor = (member: GuildMember) =>
   member.roles.cache.some((role) =>
     nitroRolesCache.some((r) => r.id === role.id)
